Entropy-based Filters

Algorithms that find ranks of importance of discrete attributes, basing on their entropy with a continous class attribute. This function is a reimplementation of FSelector's information.gain, gain.ratio and symmetrical.uncertainty.

information_gain(formula, data, x, y, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)

Arguments

formula

An object of class formula with model description.

data

A data.frame accompanying formula.

x

A data.frame or sparse matrix with attributes.

y

A vector with response variable.

type

Method name.

threads

Number of threads for parallel backend.

Value

data.frame containing the worth of attributes in the first column and their names as row names.

Details

type = "infogain" is $$H(Class) + H(Attribute) - H(Class, Attribute)$$

type = "gainratio" is $$\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute)}$$

type = "symuncert" is $$2\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute) + H(Class)}$$

where H(X) is Shannon's Entropy for a variable X and H(X, Y) is a conditional Shannon's Entropy for a variable X with a condition to Y.

Examples

irisX <- iris[-5]
y <- iris$Species

## data.frame interface
information_gain(x = irisX, y = y)

#>              importance
#> Sepal.Length  0.4521286
#> Sepal.Width   0.2672750
#> Petal.Length  0.9402853
#> Petal.Width   0.9554360

# formula interface
information_gain(formula = Species ~ ., data = iris)

#>              importance
#> Sepal.Length  0.4521286
#> Sepal.Width   0.2672750
#> Petal.Length  0.9402853
#> Petal.Width   0.9554360

information_gain(formula = Species ~ ., data = iris, type = "gainratio")

#>              importance
#> Sepal.Length  0.4196464
#> Sepal.Width   0.2472972
#> Petal.Length  0.8584937
#> Petal.Width   0.8713692

information_gain(formula = Species ~ ., data = iris, type = "symuncert")

#>              importance
#> Sepal.Length  0.4155563
#> Sepal.Width   0.2452743
#> Petal.Length  0.8571872
#> Petal.Width   0.8705214

# sparse matrix interface
library(Matrix)
i <- c(1, 3:8); j <- c(2, 9, 6:10); x <- 7 * (1:7)
x <- sparseMatrix(i, j, x = x)
y <- c(1, 1, 1, 1, 2, 2, 2, 2)

information_gain(x = x, y = y)

#>    importance
#> 1  0.00000000
#> 2  0.09560259
#> 3  0.00000000
#> 4  0.00000000
#> 5  0.00000000
#> 6  0.09560259
#> 7  0.09560259
#> 8  0.09560259
#> 9  0.17328680
#> 10 0.09560259

information_gain(x = x, y = y, type = "gainratio")

#>    importance
#> 1         NaN
#> 2   0.2537425
#> 3         NaN
#> 4         NaN
#> 5         NaN
#> 6   0.2537425
#> 7   0.2537425
#> 8   0.2537425
#> 9   0.2355650
#> 10  0.2537425

information_gain(x = x, y = y, type = "symuncert")

#>    importance
#> 1   0.0000000
#> 2   0.1787102
#> 3   0.0000000
#> 4   0.0000000
#> 5   0.0000000
#> 6   0.1787102
#> 7   0.1787102
#> 8   0.1787102
#> 9   0.2425679
#> 10  0.1787102