Discretization

Discretize a range of numeric attributes in the dataset into nominal attributes. Minimum Description Length (MDL) method is set as the default control. There is also available equalsizeControl method.

discretize(x, y, control = list(mdlControl(), equalsizeControl()),
  all = FALSE, call = NULL)

mdlControl()

equalsizeControl(k = 10)

Arguments

x

Explanatory continuous variables to be discretized or a formula.

y

Dependent variable for supervised discretization or a data.frame when x ia a formula.

control

discretizationControl object containing the parameters for discretization algorithm. Possible inputs are mdlControl or equalsizeControl, so far. If passed as a list, the first element is used.

all

Logical indicating if a returned data.frame should contain factor features that were not discretize. (Example: should Species be returned, when you pass iris and discretize all continuous features.)

call

Keep as NULL. Inner method parameter for consistency.

k

Number of partitions.

References

U. M. Fayyad and K. B. Irani. Multi-Interval Discretization of Continuous-Valued Attributes for Classification Learning. In 13th International Joint Conference on Uncertainly in Artificial Intelligence(IJCAI93), pages 1022-1029, 1993.

Examples

# vectors
discretize(x = iris[[1]], y = iris[[5]])

#>     Sepal.Length    Species
#> 1    (-Inf,5.55]     setosa
#> 2    (-Inf,5.55]     setosa
#> 3    (-Inf,5.55]     setosa
#> 4    (-Inf,5.55]     setosa
#> 5    (-Inf,5.55]     setosa
#> 6    (-Inf,5.55]     setosa
#> 7    (-Inf,5.55]     setosa
#> 8    (-Inf,5.55]     setosa
#> 9    (-Inf,5.55]     setosa
#> 10   (-Inf,5.55]     setosa
#> 11   (-Inf,5.55]     setosa
#> 12   (-Inf,5.55]     setosa
#> 13   (-Inf,5.55]     setosa
#> 14   (-Inf,5.55]     setosa
#> 15   (5.55,6.15]     setosa
#> 16   (5.55,6.15]     setosa
#> 17   (-Inf,5.55]     setosa
#> 18   (-Inf,5.55]     setosa
#> 19   (5.55,6.15]     setosa
#> 20   (-Inf,5.55]     setosa
#> 21   (-Inf,5.55]     setosa
#> 22   (-Inf,5.55]     setosa
#> 23   (-Inf,5.55]     setosa
#> 24   (-Inf,5.55]     setosa
#> 25   (-Inf,5.55]     setosa
#> 26   (-Inf,5.55]     setosa
#> 27   (-Inf,5.55]     setosa
#> 28   (-Inf,5.55]     setosa
#> 29   (-Inf,5.55]     setosa
#> 30   (-Inf,5.55]     setosa
#> 31   (-Inf,5.55]     setosa
#> 32   (-Inf,5.55]     setosa
#> 33   (-Inf,5.55]     setosa
#> 34   (-Inf,5.55]     setosa
#> 35   (-Inf,5.55]     setosa
#> 36   (-Inf,5.55]     setosa
#> 37   (-Inf,5.55]     setosa
#> 38   (-Inf,5.55]     setosa
#> 39   (-Inf,5.55]     setosa
#> 40   (-Inf,5.55]     setosa
#> 41   (-Inf,5.55]     setosa
#> 42   (-Inf,5.55]     setosa
#> 43   (-Inf,5.55]     setosa
#> 44   (-Inf,5.55]     setosa
#> 45   (-Inf,5.55]     setosa
#> 46   (-Inf,5.55]     setosa
#> 47   (-Inf,5.55]     setosa
#> 48   (-Inf,5.55]     setosa
#> 49   (-Inf,5.55]     setosa
#> 50   (-Inf,5.55]     setosa
#> 51   (6.15, Inf] versicolor
#> 52   (6.15, Inf] versicolor
#> 53   (6.15, Inf] versicolor
#> 54   (-Inf,5.55] versicolor
#> 55   (6.15, Inf] versicolor
#> 56   (5.55,6.15] versicolor
#> 57   (6.15, Inf] versicolor
#> 58   (-Inf,5.55] versicolor
#> 59   (6.15, Inf] versicolor
#> 60   (-Inf,5.55] versicolor
#> 61   (-Inf,5.55] versicolor
#> 62   (5.55,6.15] versicolor
#> 63   (5.55,6.15] versicolor
#> 64   (5.55,6.15] versicolor
#> 65   (5.55,6.15] versicolor
#> 66   (6.15, Inf] versicolor
#> 67   (5.55,6.15] versicolor
#> 68   (5.55,6.15] versicolor
#> 69   (6.15, Inf] versicolor
#> 70   (5.55,6.15] versicolor
#> 71   (5.55,6.15] versicolor
#> 72   (5.55,6.15] versicolor
#> 73   (6.15, Inf] versicolor
#> 74   (5.55,6.15] versicolor
#> 75   (6.15, Inf] versicolor
#> 76   (6.15, Inf] versicolor
#> 77   (6.15, Inf] versicolor
#> 78   (6.15, Inf] versicolor
#> 79   (5.55,6.15] versicolor
#> 80   (5.55,6.15] versicolor
#> 81   (-Inf,5.55] versicolor
#> 82   (-Inf,5.55] versicolor
#> 83   (5.55,6.15] versicolor
#> 84   (5.55,6.15] versicolor
#> 85   (-Inf,5.55] versicolor
#> 86   (5.55,6.15] versicolor
#> 87   (6.15, Inf] versicolor
#> 88   (6.15, Inf] versicolor
#> 89   (5.55,6.15] versicolor
#> 90   (-Inf,5.55] versicolor
#> 91   (-Inf,5.55] versicolor
#> 92   (5.55,6.15] versicolor
#> 93   (5.55,6.15] versicolor
#> 94   (-Inf,5.55] versicolor
#> 95   (5.55,6.15] versicolor
#> 96   (5.55,6.15] versicolor
#> 97   (5.55,6.15] versicolor
#> 98   (6.15, Inf] versicolor
#> 99   (-Inf,5.55] versicolor
#> 100  (5.55,6.15] versicolor
#> 101  (6.15, Inf]  virginica
#> 102  (5.55,6.15]  virginica
#> 103  (6.15, Inf]  virginica
#> 104  (6.15, Inf]  virginica
#> 105  (6.15, Inf]  virginica
#> 106  (6.15, Inf]  virginica
#> 107  (-Inf,5.55]  virginica
#> 108  (6.15, Inf]  virginica
#> 109  (6.15, Inf]  virginica
#> 110  (6.15, Inf]  virginica
#> 111  (6.15, Inf]  virginica
#> 112  (6.15, Inf]  virginica
#> 113  (6.15, Inf]  virginica
#> 114  (5.55,6.15]  virginica
#> 115  (5.55,6.15]  virginica
#> 116  (6.15, Inf]  virginica
#> 117  (6.15, Inf]  virginica
#> 118  (6.15, Inf]  virginica
#> 119  (6.15, Inf]  virginica
#> 120  (5.55,6.15]  virginica
#> 121  (6.15, Inf]  virginica
#> 122  (5.55,6.15]  virginica
#> 123  (6.15, Inf]  virginica
#> 124  (6.15, Inf]  virginica
#> 125  (6.15, Inf]  virginica
#> 126  (6.15, Inf]  virginica
#> 127  (6.15, Inf]  virginica
#> 128  (5.55,6.15]  virginica
#> 129  (6.15, Inf]  virginica
#> 130  (6.15, Inf]  virginica
#> 131  (6.15, Inf]  virginica
#> 132  (6.15, Inf]  virginica
#> 133  (6.15, Inf]  virginica
#> 134  (6.15, Inf]  virginica
#> 135  (5.55,6.15]  virginica
#> 136  (6.15, Inf]  virginica
#> 137  (6.15, Inf]  virginica
#> 138  (6.15, Inf]  virginica
#> 139  (5.55,6.15]  virginica
#> 140  (6.15, Inf]  virginica
#> 141  (6.15, Inf]  virginica
#> 142  (6.15, Inf]  virginica
#> 143  (5.55,6.15]  virginica
#> 144  (6.15, Inf]  virginica
#> 145  (6.15, Inf]  virginica
#> 146  (6.15, Inf]  virginica
#> 147  (6.15, Inf]  virginica
#> 148  (6.15, Inf]  virginica
#> 149  (6.15, Inf]  virginica
#> 150  (5.55,6.15]  virginica

# list and vector
discretize(x = list(iris[[1]], iris$Sepal.Width), y = iris$Species)

#>     Sepal.Length Sepal.Width    Species
#> 1    (-Inf,5.55] (3.35, Inf]     setosa
#> 2    (-Inf,5.55] (2.95,3.35]     setosa
#> 3    (-Inf,5.55] (2.95,3.35]     setosa
#> 4    (-Inf,5.55] (2.95,3.35]     setosa
#> 5    (-Inf,5.55] (3.35, Inf]     setosa
#> 6    (-Inf,5.55] (3.35, Inf]     setosa
#> 7    (-Inf,5.55] (3.35, Inf]     setosa
#> 8    (-Inf,5.55] (3.35, Inf]     setosa
#> 9    (-Inf,5.55] (-Inf,2.95]     setosa
#> 10   (-Inf,5.55] (2.95,3.35]     setosa
#> 11   (-Inf,5.55] (3.35, Inf]     setosa
#> 12   (-Inf,5.55] (3.35, Inf]     setosa
#> 13   (-Inf,5.55] (2.95,3.35]     setosa
#> 14   (-Inf,5.55] (2.95,3.35]     setosa
#> 15   (5.55,6.15] (3.35, Inf]     setosa
#> 16   (5.55,6.15] (3.35, Inf]     setosa
#> 17   (-Inf,5.55] (3.35, Inf]     setosa
#> 18   (-Inf,5.55] (3.35, Inf]     setosa
#> 19   (5.55,6.15] (3.35, Inf]     setosa
#> 20   (-Inf,5.55] (3.35, Inf]     setosa
#> 21   (-Inf,5.55] (3.35, Inf]     setosa
#> 22   (-Inf,5.55] (3.35, Inf]     setosa
#> 23   (-Inf,5.55] (3.35, Inf]     setosa
#> 24   (-Inf,5.55] (2.95,3.35]     setosa
#> 25   (-Inf,5.55] (3.35, Inf]     setosa
#> 26   (-Inf,5.55] (2.95,3.35]     setosa
#> 27   (-Inf,5.55] (3.35, Inf]     setosa
#> 28   (-Inf,5.55] (3.35, Inf]     setosa
#> 29   (-Inf,5.55] (3.35, Inf]     setosa
#> 30   (-Inf,5.55] (2.95,3.35]     setosa
#> 31   (-Inf,5.55] (2.95,3.35]     setosa
#> 32   (-Inf,5.55] (3.35, Inf]     setosa
#> 33   (-Inf,5.55] (3.35, Inf]     setosa
#> 34   (-Inf,5.55] (3.35, Inf]     setosa
#> 35   (-Inf,5.55] (2.95,3.35]     setosa
#> 36   (-Inf,5.55] (2.95,3.35]     setosa
#> 37   (-Inf,5.55] (3.35, Inf]     setosa
#> 38   (-Inf,5.55] (3.35, Inf]     setosa
#> 39   (-Inf,5.55] (2.95,3.35]     setosa
#> 40   (-Inf,5.55] (3.35, Inf]     setosa
#> 41   (-Inf,5.55] (3.35, Inf]     setosa
#> 42   (-Inf,5.55] (-Inf,2.95]     setosa
#> 43   (-Inf,5.55] (2.95,3.35]     setosa
#> 44   (-Inf,5.55] (3.35, Inf]     setosa
#> 45   (-Inf,5.55] (3.35, Inf]     setosa
#> 46   (-Inf,5.55] (2.95,3.35]     setosa
#> 47   (-Inf,5.55] (3.35, Inf]     setosa
#> 48   (-Inf,5.55] (2.95,3.35]     setosa
#> 49   (-Inf,5.55] (3.35, Inf]     setosa
#> 50   (-Inf,5.55] (2.95,3.35]     setosa
#> 51   (6.15, Inf] (2.95,3.35] versicolor
#> 52   (6.15, Inf] (2.95,3.35] versicolor
#> 53   (6.15, Inf] (2.95,3.35] versicolor
#> 54   (-Inf,5.55] (-Inf,2.95] versicolor
#> 55   (6.15, Inf] (-Inf,2.95] versicolor
#> 56   (5.55,6.15] (-Inf,2.95] versicolor
#> 57   (6.15, Inf] (2.95,3.35] versicolor
#> 58   (-Inf,5.55] (-Inf,2.95] versicolor
#> 59   (6.15, Inf] (-Inf,2.95] versicolor
#> 60   (-Inf,5.55] (-Inf,2.95] versicolor
#> 61   (-Inf,5.55] (-Inf,2.95] versicolor
#> 62   (5.55,6.15] (2.95,3.35] versicolor
#> 63   (5.55,6.15] (-Inf,2.95] versicolor
#> 64   (5.55,6.15] (-Inf,2.95] versicolor
#> 65   (5.55,6.15] (-Inf,2.95] versicolor
#> 66   (6.15, Inf] (2.95,3.35] versicolor
#> 67   (5.55,6.15] (2.95,3.35] versicolor
#> 68   (5.55,6.15] (-Inf,2.95] versicolor
#> 69   (6.15, Inf] (-Inf,2.95] versicolor
#> 70   (5.55,6.15] (-Inf,2.95] versicolor
#> 71   (5.55,6.15] (2.95,3.35] versicolor
#> 72   (5.55,6.15] (-Inf,2.95] versicolor
#> 73   (6.15, Inf] (-Inf,2.95] versicolor
#> 74   (5.55,6.15] (-Inf,2.95] versicolor
#> 75   (6.15, Inf] (-Inf,2.95] versicolor
#> 76   (6.15, Inf] (2.95,3.35] versicolor
#> 77   (6.15, Inf] (-Inf,2.95] versicolor
#> 78   (6.15, Inf] (2.95,3.35] versicolor
#> 79   (5.55,6.15] (-Inf,2.95] versicolor
#> 80   (5.55,6.15] (-Inf,2.95] versicolor
#> 81   (-Inf,5.55] (-Inf,2.95] versicolor
#> 82   (-Inf,5.55] (-Inf,2.95] versicolor
#> 83   (5.55,6.15] (-Inf,2.95] versicolor
#> 84   (5.55,6.15] (-Inf,2.95] versicolor
#> 85   (-Inf,5.55] (2.95,3.35] versicolor
#> 86   (5.55,6.15] (3.35, Inf] versicolor
#> 87   (6.15, Inf] (2.95,3.35] versicolor
#> 88   (6.15, Inf] (-Inf,2.95] versicolor
#> 89   (5.55,6.15] (2.95,3.35] versicolor
#> 90   (-Inf,5.55] (-Inf,2.95] versicolor
#> 91   (-Inf,5.55] (-Inf,2.95] versicolor
#> 92   (5.55,6.15] (2.95,3.35] versicolor
#> 93   (5.55,6.15] (-Inf,2.95] versicolor
#> 94   (-Inf,5.55] (-Inf,2.95] versicolor
#> 95   (5.55,6.15] (-Inf,2.95] versicolor
#> 96   (5.55,6.15] (2.95,3.35] versicolor
#> 97   (5.55,6.15] (-Inf,2.95] versicolor
#> 98   (6.15, Inf] (-Inf,2.95] versicolor
#> 99   (-Inf,5.55] (-Inf,2.95] versicolor
#> 100  (5.55,6.15] (-Inf,2.95] versicolor
#> 101  (6.15, Inf] (2.95,3.35]  virginica
#> 102  (5.55,6.15] (-Inf,2.95]  virginica
#> 103  (6.15, Inf] (2.95,3.35]  virginica
#> 104  (6.15, Inf] (-Inf,2.95]  virginica
#> 105  (6.15, Inf] (2.95,3.35]  virginica
#> 106  (6.15, Inf] (2.95,3.35]  virginica
#> 107  (-Inf,5.55] (-Inf,2.95]  virginica
#> 108  (6.15, Inf] (-Inf,2.95]  virginica
#> 109  (6.15, Inf] (-Inf,2.95]  virginica
#> 110  (6.15, Inf] (3.35, Inf]  virginica
#> 111  (6.15, Inf] (2.95,3.35]  virginica
#> 112  (6.15, Inf] (-Inf,2.95]  virginica
#> 113  (6.15, Inf] (2.95,3.35]  virginica
#> 114  (5.55,6.15] (-Inf,2.95]  virginica
#> 115  (5.55,6.15] (-Inf,2.95]  virginica
#> 116  (6.15, Inf] (2.95,3.35]  virginica
#> 117  (6.15, Inf] (2.95,3.35]  virginica
#> 118  (6.15, Inf] (3.35, Inf]  virginica
#> 119  (6.15, Inf] (-Inf,2.95]  virginica
#> 120  (5.55,6.15] (-Inf,2.95]  virginica
#> 121  (6.15, Inf] (2.95,3.35]  virginica
#> 122  (5.55,6.15] (-Inf,2.95]  virginica
#> 123  (6.15, Inf] (-Inf,2.95]  virginica
#> 124  (6.15, Inf] (-Inf,2.95]  virginica
#> 125  (6.15, Inf] (2.95,3.35]  virginica
#> 126  (6.15, Inf] (2.95,3.35]  virginica
#> 127  (6.15, Inf] (-Inf,2.95]  virginica
#> 128  (5.55,6.15] (2.95,3.35]  virginica
#> 129  (6.15, Inf] (-Inf,2.95]  virginica
#> 130  (6.15, Inf] (2.95,3.35]  virginica
#> 131  (6.15, Inf] (-Inf,2.95]  virginica
#> 132  (6.15, Inf] (3.35, Inf]  virginica
#> 133  (6.15, Inf] (-Inf,2.95]  virginica
#> 134  (6.15, Inf] (-Inf,2.95]  virginica
#> 135  (5.55,6.15] (-Inf,2.95]  virginica
#> 136  (6.15, Inf] (2.95,3.35]  virginica
#> 137  (6.15, Inf] (3.35, Inf]  virginica
#> 138  (6.15, Inf] (2.95,3.35]  virginica
#> 139  (5.55,6.15] (2.95,3.35]  virginica
#> 140  (6.15, Inf] (2.95,3.35]  virginica
#> 141  (6.15, Inf] (2.95,3.35]  virginica
#> 142  (6.15, Inf] (2.95,3.35]  virginica
#> 143  (5.55,6.15] (-Inf,2.95]  virginica
#> 144  (6.15, Inf] (2.95,3.35]  virginica
#> 145  (6.15, Inf] (2.95,3.35]  virginica
#> 146  (6.15, Inf] (2.95,3.35]  virginica
#> 147  (6.15, Inf] (-Inf,2.95]  virginica
#> 148  (6.15, Inf] (2.95,3.35]  virginica
#> 149  (6.15, Inf] (3.35, Inf]  virginica
#> 150  (5.55,6.15] (2.95,3.35]  virginica

# formula input
discretize(x = Species ~ ., y = iris)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 2    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 3    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 4    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 5    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 6    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 7    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 8    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 9    (-Inf,5.55] (-Inf,2.95]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 10   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 11   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 12   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 13   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 14   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 15   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 16   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 17   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 18   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 19   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 20   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 21   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 22   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 23   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 24   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 25   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 26   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 27   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 28   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 29   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 30   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 31   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 32   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 33   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 34   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 35   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 36   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 37   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 38   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 39   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 40   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 41   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 42   (-Inf,5.55] (-Inf,2.95]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 43   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 44   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 45   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 46   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 47   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 48   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 49   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 50   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 51   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 52   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 53   (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75] versicolor
#> 54   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 55   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 56   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 57   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 58   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 59   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 60   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 61   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 62   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 63   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 64   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 65   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 66   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 67   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 68   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 69   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 70   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 71   (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf] versicolor
#> 72   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 73   (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 74   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 75   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 76   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 77   (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 78   (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75] versicolor
#> 79   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 80   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 81   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 82   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 83   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 84   (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 85   (-Inf,5.55] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 86   (5.55,6.15] (3.35, Inf]  (2.45,4.75]  (0.8,1.75] versicolor
#> 87   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 88   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 89   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 90   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 91   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 92   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 93   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 94   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 95   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 96   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 97   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 98   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 99   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 100  (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 101  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 102  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 103  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 104  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 105  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 106  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 107  (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75]  virginica
#> 108  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 109  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 110  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 111  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 112  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 113  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 114  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 115  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 116  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 117  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 118  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 119  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 120  (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 121  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 122  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 123  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 124  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 125  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 126  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 127  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 128  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 129  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 130  (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75]  virginica
#> 131  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 132  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 133  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 134  (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 135  (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 136  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 137  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 138  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 139  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 140  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 141  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 142  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 143  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 144  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 145  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 146  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 147  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 148  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 149  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 150  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica

discretize(Species ~ ., iris)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 2    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 3    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 4    (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 5    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 6    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 7    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 8    (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 9    (-Inf,5.55] (-Inf,2.95]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 10   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 11   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 12   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 13   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 14   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 15   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 16   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 17   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 18   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 19   (5.55,6.15] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 20   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 21   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 22   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 23   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 24   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 25   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 26   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 27   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 28   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 29   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 30   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 31   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 32   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 33   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 34   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 35   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 36   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 37   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 38   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 39   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 40   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 41   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 42   (-Inf,5.55] (-Inf,2.95]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 43   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 44   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 45   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 46   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 47   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 48   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 49   (-Inf,5.55] (3.35, Inf]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 50   (-Inf,5.55] (2.95,3.35]  (-Inf,2.45]  (-Inf,0.8]     setosa
#> 51   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 52   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 53   (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75] versicolor
#> 54   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 55   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 56   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 57   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 58   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 59   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 60   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 61   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 62   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 63   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 64   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 65   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 66   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 67   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 68   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 69   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 70   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 71   (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf] versicolor
#> 72   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 73   (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 74   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 75   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 76   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 77   (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 78   (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75] versicolor
#> 79   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 80   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 81   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 82   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 83   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 84   (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75] versicolor
#> 85   (-Inf,5.55] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 86   (5.55,6.15] (3.35, Inf]  (2.45,4.75]  (0.8,1.75] versicolor
#> 87   (6.15, Inf] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 88   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 89   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 90   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 91   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 92   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 93   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 94   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 95   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 96   (5.55,6.15] (2.95,3.35]  (2.45,4.75]  (0.8,1.75] versicolor
#> 97   (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 98   (6.15, Inf] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 99   (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 100  (5.55,6.15] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75] versicolor
#> 101  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 102  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 103  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 104  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 105  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 106  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 107  (-Inf,5.55] (-Inf,2.95]  (2.45,4.75]  (0.8,1.75]  virginica
#> 108  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 109  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 110  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 111  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 112  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 113  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 114  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 115  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 116  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 117  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 118  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 119  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 120  (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 121  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 122  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 123  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 124  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 125  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 126  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 127  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 128  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 129  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 130  (6.15, Inf] (2.95,3.35]  (4.75, Inf]  (0.8,1.75]  virginica
#> 131  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 132  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 133  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 134  (6.15, Inf] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 135  (5.55,6.15] (-Inf,2.95]  (4.75, Inf]  (0.8,1.75]  virginica
#> 136  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 137  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 138  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 139  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 140  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 141  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 142  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 143  (5.55,6.15] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 144  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 145  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 146  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 147  (6.15, Inf] (-Inf,2.95]  (4.75, Inf] (1.75, Inf]  virginica
#> 148  (6.15, Inf] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica
#> 149  (6.15, Inf] (3.35, Inf]  (4.75, Inf] (1.75, Inf]  virginica
#> 150  (5.55,6.15] (2.95,3.35]  (4.75, Inf] (1.75, Inf]  virginica

## Not run: ------------------------------------
# # Same results
# library(RWeka)
# Rweka_disc_out <- RWeka::Discretize(Species ~ Sepal.Length, iris)[, 1]
# FSelectorRcpp_disc_out <- FSelectorRcpp::discretize(Species ~ Sepal.Length,
#                                                     iris)[, 1]
# table(Rweka_disc_out, FSelectorRcpp_disc_out)
# # But faster method
# library(microbenchmark)
# microbenchmark(FSelectorRcpp::discretize(Species ~ Sepal.Length, iris),
#                RWeka::Discretize(Species ~ Sepal.Length, iris))
# 
## ---------------------------------------------