General Feature Searching Engine

A convenience wrapper for greedy and exhaustive feature selection algorithms that extract valuable attributes depending on the evaluation method (called evaluator).

feature_search(attributes, fun, data, mode = c("greedy", "exhaustive"),
  type = c("forward", "backward"), sizes = 1:length(attributes),
  parallel = TRUE, ...)

Arguments

attributes

A character vector with attributes' names to be used to extract the most valuable features.

fun

A function (evaluator) to be used to score features' sets at each iteration of the algorithm passed via mode. See Examples.

data

A data set for fun function (evaluator).

mode

A character that determines which search algorithm to perform.

type

Used when mode = "greedy" - whether to use the backward or the forward multiple-way search.

sizes

Used when mode = "exhaustive" - a vector of sizes of attributes subsets.

parallel

Allow parallelization.

…

Other arguments passed to foreach function.

Value

A list with following components

best - a data.frame with the best subset and it's score (1 - feature used, 0 - feature not used),

all - a data.frame with all checked features' subsets and their score (1 - feature used, 0 - feature not used),

data - the data used in the feature selection,

fun - the evaluator used to compute the score of importance for features' subsets,

call - an origin call of the feature_search,

mode - the mode used in the call.

Details

The evaluator function passed with fun is used to determine the importance score of current features' subset. The score is used in a multiple-way (backward or forward) greedy algorithm as a stopping moment or as a selection criterion in the exhaustive search that checks all possible attributes' subset combinations (of sizes passed in sizes).

Note

Note that score depends on the evaluator you provide in the fun parameter.

Examples

# Enable parallelization in examples
 library(doSNOW) # doSNOW has an option for progress bar
 cl <- makeCluster(2)
 registerDoSNOW(cl)

# Close at the end
# stopCluster(cl)
# registerDoSEQ()

# 1) Evaluator from FSelector package.
evaluator <- function(subset, data, dependent = names(iris)[5]) {
  library(rpart)
  k <- 5
  splits <- runif(nrow(data))
  results <- sapply(1:k, function(i) {
    test.idx <- (splits >= (i - 1) / k) & (splits < i / k)
    train.idx <- !test.idx
    test <- data[test.idx, , drop = FALSE]
    train <- data[train.idx, , drop = FALSE]
    tree <- rpart(to_formula(subset, dependent), train)
    error.rate <- sum(test[[dependent]] != predict(tree, test, type = "c")) /
    nrow(test)
    return(1 - error.rate)
  })
  return(mean(results))
}

# Default greedy search.
system.time(
  feature_search(attributes = names(iris)[-5],
                 fun = evaluator,
                 data = iris)
)

#>   |                                                                                                                            |                                                                                                                    |   0%  |                                                                                                                            |=============================                                                                                       |  25%  |                                                                                                                            |==========================================================                                                          |  50%  |                                                                                                                            |=======================================================================================                             |  75%  |                                                                                                                            |====================================================================================================================| 100%  |                                                                                                                              |                                                                                                                      |   0%  |                                                                                                                              |=======================================                                                                               |  33%  |                                                                                                                              |===============================================================================                                       |  67%  |                                                                                                                              |======================================================================================================================| 100%  |                                                                                                                              |                                                                                                                      |   0%  |                                                                                                                              |===========================================================                                                           |  50%  |                                                                                                                              |======================================================================================================================| 100%

#>    user  system elapsed 
#>   0.962   0.693   0.496 

system.time(
  feature_search(attributes = names(iris)[-5],
                 fun = evaluator,
                 data = iris,
                 parallel = FALSE)
)

#>    user  system elapsed 
#>   0.215   0.004   0.219 

# Optional exhaustive search.
system.time(
  feature_search(attributes = names(iris)[-5],
                 fun = evaluator,
                 data = iris,
                 mode = "exhaustive")
)

#>   |                                                                                                                              |                                                                                                                      |   0%  |                                                                                                                              |========                                                                                                              |   7%  |                                                                                                                              |================                                                                                                      |  13%  |                                                                                                                              |========================                                                                                              |  20%  |                                                                                                                              |===============================                                                                                       |  27%  |                                                                                                                              |=======================================                                                                               |  33%  |                                                                                                                              |===============================================                                                                       |  40%  |                                                                                                                              |=======================================================                                                               |  47%  |                                                                                                                              |===============================================================                                                       |  53%  |                                                                                                                              |=======================================================================                                               |  60%  |                                                                                                                              |===============================================================================                                       |  67%  |                                                                                                                              |=======================================================================================                               |  73%  |                                                                                                                              |==============================================================================================                        |  80%  |                                                                                                                              |======================================================================================================                |  87%  |                                                                                                                              |==============================================================================================================        |  93%  |                                                                                                                              |======================================================================================================================| 100%

#>    user  system elapsed 
#>   0.016   0.001   0.248 

system.time(
  feature_search(attributes = names(iris)[-5],
                 fun = evaluator,
                 data = iris,
                 mode = "exhaustive",
                 parallel = FALSE)
)

#>    user  system elapsed 
#>   0.312   0.000   0.311 

# 2) Maximize R^2 statistics in the linear regression model/problem.

evaluator_R2_lm <- function(attributes, data, dependent = names(iris)[1]) {
  summary(
    lm(to_formula(attributes, dependent), data = data)
  )$r.squared
}

feature_search(attributes = names(iris)[-1],
               fun = evaluator_R2_lm, data = iris,
               mode = "exhaustive")

#>   |                                                                                                                              |                                                                                                                      |   0%  |                                                                                                                              |========                                                                                                              |   7%  |                                                                                                                              |================                                                                                                      |  13%  |                                                                                                                              |========================                                                                                              |  20%  |                                                                                                                              |===============================                                                                                       |  27%  |                                                                                                                              |=======================================                                                                               |  33%  |                                                                                                                              |===============================================                                                                       |  40%  |                                                                                                                              |=======================================================                                                               |  47%  |                                                                                                                              |===============================================================                                                       |  53%  |                                                                                                                              |=======================================================================                                               |  60%  |                                                                                                                              |===============================================================================                                       |  67%  |                                                                                                                              |=======================================================================================                               |  73%  |                                                                                                                              |==============================================================================================                        |  80%  |                                                                                                                              |======================================================================================================                |  87%  |                                                                                                                              |==============================================================================================================        |  93%  |                                                                                                                              |======================================================================================================================| 100%

#> $best
#>    Sepal.Width Petal.Length Petal.Width Species    values
#> 15           1            1           1       1 0.8673123
#> 
#> $all
#>    Sepal.Width Petal.Length Petal.Width Species     values
#> 1            1            0           0       0 0.01382265
#> 2            0            1           0       0  0.7599546
#> 3            0            0           1       0  0.6690277
#> 4            0            0           0       1  0.6187057
#> 5            1            1           0       0  0.8401778
#> 6            1            0           1       0  0.7072371
#> 7            1            0           0       1  0.7259066
#> 8            0            1           1       0  0.7662613
#> 9            0            1           0       1  0.8367238
#> 10           0            0           1       1  0.6693664
#> 11           1            1           1       0  0.8586117
#> 12           1            1           0       1  0.8633088
#> 13           1            0           1       1  0.7323845
#> 14           0            1           1       1  0.8367254
#> 15           1            1           1       1  0.8673123
#> 
#> $fun
#> function (attributes, data, dependent = names(iris)[1]) 
#> {
#>     summary(lm(to_formula(attributes, dependent), data = data))$r.squared
#> }
#> <environment: 0xe7fb818>
#> 
#> $call
#> feature_search(attributes = names(iris)[-1], fun = evaluator_R2_lm, 
#>     data = iris, mode = "exhaustive")
#> 
#> $mode
#> [1] "exhaustive"
#> 
#> $data.Sepal.Length
#>   [1] 5.1 4.9 4.7 4.6 5.0 5.4 4.6 5.0 4.4 4.9 5.4 4.8 4.8 4.3 5.8 5.7 5.4 5.1 5.7 5.1 5.4 5.1 4.6 5.1 4.8 5.0 5.0 5.2 5.2 4.7
#>  [31] 4.8 5.4 5.2 5.5 4.9 5.0 5.5 4.9 4.4 5.1 5.0 4.5 4.4 5.0 5.1 4.8 5.1 4.6 5.3 5.0 7.0 6.4 6.9 5.5 6.5 5.7 6.3 4.9 6.6 5.2
#>  [61] 5.0 5.9 6.0 6.1 5.6 6.7 5.6 5.8 6.2 5.6 5.9 6.1 6.3 6.1 6.4 6.6 6.8 6.7 6.0 5.7 5.5 5.5 5.8 6.0 5.4 6.0 6.7 6.3 5.6 5.5
#>  [91] 5.5 6.1 5.8 5.0 5.6 5.7 5.7 6.2 5.1 5.7 6.3 5.8 7.1 6.3 6.5 7.6 4.9 7.3 6.7 7.2 6.5 6.4 6.8 5.7 5.8 6.4 6.5 7.7 7.7 6.0
#> [121] 6.9 5.6 7.7 6.3 6.7 7.2 6.2 6.1 6.4 7.2 7.4 7.9 6.4 6.3 6.1 7.7 6.3 6.4 6.0 6.9 6.7 6.9 5.8 6.8 6.7 6.7 6.3 6.5 6.2 5.9
#> 
#> $data.Sepal.Width
#>   [1] 3.5 3.0 3.2 3.1 3.6 3.9 3.4 3.4 2.9 3.1 3.7 3.4 3.0 3.0 4.0 4.4 3.9 3.5 3.8 3.8 3.4 3.7 3.6 3.3 3.4 3.0 3.4 3.5 3.4 3.2
#>  [31] 3.1 3.4 4.1 4.2 3.1 3.2 3.5 3.6 3.0 3.4 3.5 2.3 3.2 3.5 3.8 3.0 3.8 3.2 3.7 3.3 3.2 3.2 3.1 2.3 2.8 2.8 3.3 2.4 2.9 2.7
#>  [61] 2.0 3.0 2.2 2.9 2.9 3.1 3.0 2.7 2.2 2.5 3.2 2.8 2.5 2.8 2.9 3.0 2.8 3.0 2.9 2.6 2.4 2.4 2.7 2.7 3.0 3.4 3.1 2.3 3.0 2.5
#>  [91] 2.6 3.0 2.6 2.3 2.7 3.0 2.9 2.9 2.5 2.8 3.3 2.7 3.0 2.9 3.0 3.0 2.5 2.9 2.5 3.6 3.2 2.7 3.0 2.5 2.8 3.2 3.0 3.8 2.6 2.2
#> [121] 3.2 2.8 2.8 2.7 3.3 3.2 2.8 3.0 2.8 3.0 2.8 3.8 2.8 2.8 2.6 3.0 3.4 3.1 3.0 3.1 3.1 3.1 2.7 3.2 3.3 3.0 2.5 3.0 3.4 3.0
#> 
#> $data.Petal.Length
#>   [1] 1.4 1.4 1.3 1.5 1.4 1.7 1.4 1.5 1.4 1.5 1.5 1.6 1.4 1.1 1.2 1.5 1.3 1.4 1.7 1.5 1.7 1.5 1.0 1.7 1.9 1.6 1.6 1.5 1.4 1.6
#>  [31] 1.6 1.5 1.5 1.4 1.5 1.2 1.3 1.4 1.3 1.5 1.3 1.3 1.3 1.6 1.9 1.4 1.6 1.4 1.5 1.4 4.7 4.5 4.9 4.0 4.6 4.5 4.7 3.3 4.6 3.9
#>  [61] 3.5 4.2 4.0 4.7 3.6 4.4 4.5 4.1 4.5 3.9 4.8 4.0 4.9 4.7 4.3 4.4 4.8 5.0 4.5 3.5 3.8 3.7 3.9 5.1 4.5 4.5 4.7 4.4 4.1 4.0
#>  [91] 4.4 4.6 4.0 3.3 4.2 4.2 4.2 4.3 3.0 4.1 6.0 5.1 5.9 5.6 5.8 6.6 4.5 6.3 5.8 6.1 5.1 5.3 5.5 5.0 5.1 5.3 5.5 6.7 6.9 5.0
#> [121] 5.7 4.9 6.7 4.9 5.7 6.0 4.8 4.9 5.6 5.8 6.1 6.4 5.6 5.1 5.6 6.1 5.6 5.5 4.8 5.4 5.6 5.1 5.1 5.9 5.7 5.2 5.0 5.2 5.4 5.1
#> 
#> $data.Petal.Width
#>   [1] 0.2 0.2 0.2 0.2 0.2 0.4 0.3 0.2 0.2 0.1 0.2 0.2 0.1 0.1 0.2 0.4 0.4 0.3 0.3 0.3 0.2 0.4 0.2 0.5 0.2 0.2 0.4 0.2 0.2 0.2
#>  [31] 0.2 0.4 0.1 0.2 0.2 0.2 0.2 0.1 0.2 0.2 0.3 0.3 0.2 0.6 0.4 0.3 0.2 0.2 0.2 0.2 1.4 1.5 1.5 1.3 1.5 1.3 1.6 1.0 1.3 1.4
#>  [61] 1.0 1.5 1.0 1.4 1.3 1.4 1.5 1.0 1.5 1.1 1.8 1.3 1.5 1.2 1.3 1.4 1.4 1.7 1.5 1.0 1.1 1.0 1.2 1.6 1.5 1.6 1.5 1.3 1.3 1.3
#>  [91] 1.2 1.4 1.2 1.0 1.3 1.2 1.3 1.3 1.1 1.3 2.5 1.9 2.1 1.8 2.2 2.1 1.7 1.8 1.8 2.5 2.0 1.9 2.1 2.0 2.4 2.3 1.8 2.2 2.3 1.5
#> [121] 2.3 2.0 2.0 1.8 2.1 1.8 1.8 1.8 2.1 1.6 1.9 2.0 2.2 1.5 1.4 2.3 2.4 1.8 1.8 2.1 2.4 2.3 1.9 2.3 2.5 2.3 1.9 2.0 2.3 1.8
#> 
#> $data.Species
#>   [1] setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa    
#>  [12] setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa    
#>  [23] setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa    
#>  [34] setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa     setosa    
#>  [45] setosa     setosa     setosa     setosa     setosa     setosa     versicolor versicolor versicolor versicolor versicolor
#>  [56] versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor
#>  [67] versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor
#>  [78] versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor
#>  [89] versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor versicolor
#> [100] versicolor virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#> [111] virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#> [122] virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#> [133] virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#> [144] virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#> Levels: setosa versicolor virginica
#> 

# 3) Optimize BIC crietion in generalized linear model.
# Aim of Bayesian approach it to identify the model with the highest
# probability of being the true model. - Kuha 2004

utils::data(anorexia, package = "MASS")

evaluator_BIC_glm <- function(attributes, data, dependent = "Postwt") {
  extractAIC(
    fit = glm(to_formula(attributes, dependent), family = gaussian,
              data = data),
    k = log(nrow(data))
  )[2]
}

feature_search(attributes = c("Prewt", "Treat", "offset(Prewt)"),
               fun = evaluator_BIC_glm,
               data = anorexia,
               mode = "exhaustive")

#>   |                                                                                                                              |                                                                                                                      |   0%  |                                                                                                                              |=================                                                                                                     |  14%  |                                                                                                                              |==================================                                                                                    |  29%  |                                                                                                                              |===================================================                                                                   |  43%  |                                                                                                                              |===================================================================                                                   |  57%  |                                                                                                                              |====================================================================================                                  |  71%  |                                                                                                                              |=====================================================================================================                 |  86%  |                                                                                                                              |======================================================================================================================| 100%

#> $best
#>   Prewt Treat offset(Prewt)   values
#> 3     0     0             1 508.7408
#> 
#> $all
#>   Prewt Treat offset(Prewt)   values
#> 1     1     0             0 505.5144
#> 2     0     1             0 502.1123
#> 3     0     0             1 508.7408
#> 4     1     1             0   499.08
#> 5     1     0             1 505.5144
#> 6     0     1             1 506.7839
#> 7     1     1             1   499.08
#> 
#> $fun
#> function (attributes, data, dependent = "Postwt") 
#> {
#>     extractAIC(fit = glm(to_formula(attributes, dependent), family = gaussian, 
#>         data = data), k = log(nrow(data)))[2]
#> }
#> <environment: 0xe7fb818>
#> 
#> $call
#> feature_search(attributes = c("Prewt", "Treat", "offset(Prewt)"), 
#>     fun = evaluator_BIC_glm, data = anorexia, mode = "exhaustive")
#> 
#> $mode
#> [1] "exhaustive"
#> 
#> $data.Treat
#>  [1] Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont Cont
#> [25] Cont Cont CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT  CBT 
#> [49] CBT  CBT  CBT  CBT  CBT  CBT  CBT  FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT   FT  
#> Levels: CBT Cont FT
#> 
#> $data.Prewt
#>  [1] 80.7 89.4 91.8 74.0 78.1 88.3 87.3 75.1 80.6 78.4 77.6 88.7 81.3 78.1 70.5 77.3 85.2 86.0 84.1 79.7 85.5 84.4 79.6 77.5
#> [25] 72.3 89.0 80.5 84.9 81.5 82.6 79.9 88.7 94.9 76.3 81.0 80.5 85.0 89.2 81.3 76.5 70.0 80.4 83.3 83.0 87.7 84.2 86.4 76.5
#> [49] 80.2 87.8 83.3 79.7 84.5 80.8 87.4 83.8 83.3 86.0 82.5 86.7 79.6 76.9 94.2 73.4 80.5 81.6 82.1 77.6 83.5 89.9 86.0 87.3
#> 
#> $data.Postwt
#>  [1]  80.2  80.1  86.4  86.3  76.1  78.1  75.1  86.7  73.5  84.6  77.4  79.5  89.6  81.4  81.8  77.3  84.2  75.4  79.5  73.0
#> [21]  88.3  84.7  81.4  81.2  88.2  78.8  82.2  85.6  81.4  81.9  76.4 103.6  98.4  93.4  73.4  82.1  96.7  95.3  82.4  72.5
#> [41]  90.9  71.3  85.4  81.6  89.1  83.9  82.7  75.7  82.6 100.4  85.2  83.6  84.6  96.2  86.7  95.2  94.3  91.5  91.9 100.3
#> [61]  76.7  76.8 101.6  94.9  75.2  77.8  95.5  90.7  92.5  93.8  91.7  98.0
#> 

# Close parallelization
stopCluster(cl)
registerDoSEQ()