Entropy-based Filters

Algorithms that find ranks of importance of discrete attributes, basing on their entropy with a continous class attribute. This function is a reimplementation of FSelector's information.gain, gain.ratio and symmetrical.uncertainty.

information_gain(formula, data, x, y, type = c("infogain", "gainratio",
  "symuncert"), equal = FALSE, discIntegers = FALSE, threads = 1)

Arguments

formula
An object of class formula with model description.

data
A data.frame accompanying formula.

x
A data.frame or sparse matrix with attributes.

y
A vector with response variable.

type
Method name.

equal
A logical. Whether to discretize dependent variable with the equal frequency binning discretization or not.

discIntegers
logical value. If true, then integers are treated as numeric vectors and they are discretized. If false (default) integers are treated as factors and they are left as is.

threads
Number of threads for parallel backend.

Value

data.frame with the following columns:

attributes - variables names.

importance - worth of the attributes.

Details

type = "infogain" is $$H(Class) + H(Attribute) - H(Class, Attribute)$$

type = "gainratio" is $$\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute)}$$

type = "symuncert" is $$2\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute) + H(Class)}$$

where H(X) is Shannon's Entropy for a variable X and H(X, Y) is a conditional Shannon's Entropy for a variable X with a condition to Y.

Examples

irisX <- iris[-5]
y <- iris$Species

## data.frame interface
information_gain(x = irisX, y = y)

#>     attributes importance
#> 1 Sepal.Length  0.4521286
#> 2  Sepal.Width  0.2672750
#> 3 Petal.Length  0.9402853
#> 4  Petal.Width  0.9554360

# formula interface
information_gain(formula = Species ~ ., data = iris)

#>     attributes importance
#> 1 Sepal.Length  0.4521286
#> 2  Sepal.Width  0.2672750
#> 3 Petal.Length  0.9402853
#> 4  Petal.Width  0.9554360

information_gain(formula = Species ~ ., data = iris, type = "gainratio")

#>     attributes importance
#> 1 Sepal.Length  0.4196464
#> 2  Sepal.Width  0.2472972
#> 3 Petal.Length  0.8584937
#> 4  Petal.Width  0.8713692

information_gain(formula = Species ~ ., data = iris, type = "symuncert")

#>     attributes importance
#> 1 Sepal.Length  0.4155563
#> 2  Sepal.Width  0.2452743
#> 3 Petal.Length  0.8571872
#> 4  Petal.Width  0.8705214

# sparse matrix interface
library(Matrix)
i <- c(1, 3:8); j <- c(2, 9, 6:10); x <- 7 * (1:7)
x <- sparseMatrix(i, j, x = x)
y <- c(1, 1, 1, 1, 2, 2, 2, 2)

information_gain(x = x, y = y)

#>    attributes importance
#> 1           1 0.00000000
#> 2           2 0.09560259
#> 3           3 0.00000000
#> 4           4 0.00000000
#> 5           5 0.00000000
#> 6           6 0.09560259
#> 7           7 0.09560259
#> 8           8 0.09560259
#> 9           9 0.17328680
#> 10         10 0.09560259

information_gain(x = x, y = y, type = "gainratio")

#>    attributes importance
#> 1           1        NaN
#> 2           2  0.2537425
#> 3           3        NaN
#> 4           4        NaN
#> 5           5        NaN
#> 6           6  0.2537425
#> 7           7  0.2537425
#> 8           8  0.2537425
#> 9           9  0.2355650
#> 10         10  0.2537425

information_gain(x = x, y = y, type = "symuncert")

#>    attributes importance
#> 1           1  0.0000000
#> 2           2  0.1787102
#> 3           3  0.0000000
#> 4           4  0.0000000
#> 5           5  0.0000000
#> 6           6  0.1787102
#> 7           7  0.1787102
#> 8           8  0.1787102
#> 9           9  0.2425679
#> 10         10  0.1787102